move common code into utils
